Move CI detector class to test. env package
