+++ datastore/src/test/java/io/spine/server/storage/datastore/given/TestEnvironment.java
@@ -18,14 +18,14 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.storage.datastore;
+package io.spine.server.storage.datastore.given;
 
 import static java.lang.System.getenv;
 
 /**
  * A utility class for analyzing the test environment at runtime.
  */
-final class TestEnvironment {
+public final class TestEnvironment {
 
     private static final String TRUE = "true";
